refactor:: 폴더 정리 및 게시판 media query 별로 어느 정도는 보일 수 있게 수정
