+++ src/components/community/Boards/BoardPostPreview.tsx
