+++ src/components/community/HotBoard/HotBoardPreview.tsx
@@ -1,5 +1,5 @@
 import { useGetHotPostPreview } from '@/api/hooks/community'
-import BoardPreview from '@/components/community/BoardPreview'
+import BoardPreview from '@/components/community/Boards/BoardPreview'
 
 const HotBoardPreview = () => {
   const { data: hotPosts } = useGetHotPostPreview()
